Fixing code files, names and more
